requirements.txt
python-telegram-bot==20.7
deepseek==0.8.2
python-dotenv==1.0.0